Reorganised clustering code.
